requirements.txt
Babel==0.9.6
Brlapi==0.5.6
Chameleon==2.18
GnuPGInterface==0.3.2
Jinja2==2.6
Mako==1.0.0
MarkupSafe==0.15
PAM==0.4.2
PIL==1.1.7
PasteDeploy==1.5.2
PyChart==1.39
PyOpenGL==3.0.1b2
PyWebDAV==0.9.4.1
PyYAML==3.10
Pygments==1.4
SQLAlchemy==0.9.8
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
WebOb==1.4
-e git+https://github.com/anderson89marques/WebResultadoSimulacaoPyramid.git@7af17ad518175f2b2758eaaf3e259ee2d3efbff0#egg=WebResultadosSimulacaoPyramid-master
Werkzeug==0.9.4
ZSI==2.1-a1
adium-theme-ubuntu==0.3.2
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
bzr==2.5.1
chardet==2.0.1
command-not-found==0.2.44
configglue==1.0
configobj==4.7.2
debtagshw==0.1
defer==1.0.2
deluge==1.3.11
dirspec==3.0.0
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
docutils==0.8.1
duplicity==0.6.18
feedparser==5.1
gdata==2.0.14
httplib2==0.7.2
jockey==0.9.7
keyring==0.9.2
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
lockfile==0.8
louis==2.3.0
lxml==2.3.2
mercurial==2.0.2
meutcc==0.0
mock==0.7.2
numpy==1.6.1
nvidia-common==0.0.0
oauth==1.0.1
onboard==0.97.1
oneconf==0.2.8.1
paramiko==1.7.7.1
pexpect==2.3
piston-mini-client==0.7.2
protobuf==2.4.1
psutil==0.4.1
psycopg2==2.5.4
pyOpenSSL==0.12
pycrypto==2.4.1
pycups==1.9.61
pycurl==7.19.0
pydot==1.0.2
pygame==1.9.1release
pyinotify==0.9.2
pyparsing==1.5.2
pyramid==1.5.2
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.2.2
pyramid-jinja2==2.3.3
pyramid-mako==1.0.2
pyramid-tm==0.8
pyserial==2.5
pysmbc==1.0.13
pysqlite==1.0.1
python-apt==0.8.3ubuntu7.2
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-ldap==2.3.13
python-libtorrent==0.16.17
python-openid==2.2.5
python-virtkey==0.60.0
pytz==2011k
pyxdg==0.19
reportlab==2.5
repoze.lru==0.6
rhythmbox-ubuntuone==4.2.0
sessioninstaller==0.0.0
setproctitle==1.0.1
simplejson==2.3.2
software-center-aptd-plugins==0.0.0
suds==0.4.1
system-service==0.1.6
tcc==0.0
tccv2==0.0
transaction==1.4.3
translationstring==1.3
uTidylib==0.2
ubuntuone-couch==0.3.0
ubuntuone-installer==3.0.2
ubuntuone-storage-protocol==3.0.2
ufw==0.31.1-1
unattended-upgrades==0.1
unittest2==0.5.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
urlgrabber==3.9.1
usb-creator==0.2.23
vatnumber==1.0
vboxapi==1.0
venusian==1.0
virtualenv==1.11.6
vobject==0.8.1c
wadllib==1.3.0
waitress==0.8.9
wsgiref==0.1.2
xdiagnose==2.5.3
xkit==0.0.0
xlwt==0.7.2
yum-metadata-parser==1.1.2
zope.deprecation==4.1.1
zope.interface==4.1.1
zope.sqlalchemy==0.7.5
